Calculate Non-Compartmental Analysis (NCA) Parameters

This function performs Non-Compartmental Analysis (NCA) to derive key pharmacokinetic parameters such as Cmax, AUCinf, and AUClast for specified time intervals.

calc_NCA(
  run_dir,
  ngroups,
  reference_groups,
  test_groups,
  NCA_end_time,
  samp_size
)

Arguments

run_dir

Character string specifying the path to the run directory.

ngroups

Integer specifying the total number of groups (e.g., 4 for an ABBA design).

reference_groups

Numeric vector indicating the group IDs that are designated as reference.

test_groups

Numeric vector indicating the group IDs that are designated as test.

NCA_end_time

Numeric value specifying the end time for calculations of AUClast and AUCinf.

samp_size

Integer indicating the sample size or the total number of samples for the analysis.

Value

The function returns a list containing the derived NCA parameters for each sample.

Details

The calc_NCA function internally calls getNCA for each sample in the sequence from 1 to samp_size. Note that the function is currently executed in a serial manner and is not parallelized.

Examples

if (FALSE) {
run_dir <- "c:/Workspace/mbbe"
ngroups <- 4
reference_groups <- c(1,2)
test_groups <- c(3,4)
NCA_end_time <- 7
samp_size <- 6
calc_NCA(run_dir, ngroups, reference_groups, test_groups, NCA_end_time, samp_size)
}