Calculate Bioequivalence Power

Computes the power for bioequivalence (BE) testing based on EMA standards statistics applied to each Monte Carlo simulation. The power is determined by the proportion of simulations that meet the BE criteria.

calc_power(run_dir, samp_size, alpha, model_averaging_by, NTID)

Arguments

run_dir

Character string specifying the run directory where simulation outputs are located.

samp_size

Integer indicating the number of samples to be used in the analysis.

alpha

Numeric value representing the alpha error rate. It must lie between 0 and 1.

model_averaging_by

Character string indicating the method for model averaging, either "subject" or "study".

NTID

Logical indicating if the drug being tested is a narrow therapeutic index drug.

Value

A list containing the results for:

Cmax_result: Power for the Cmax parameter.

AUCinf_result: Power for the AUCinf parameter.

AUClast_result: Power for the AUClast parameter. All power values range between 0 and 1.

Details

When the simulation is conducted by study (i.e., a unique model for each study), this results in model averaging at the study level. If model_averaging_by is set to "subject", data from different studies are merged. For each study dataset, subjects are randomly selected (without replacement) from across all studies.

The function iterates over each sample, reading the corresponding NCAresults (designated by the sample number). Subsequently, it determines if each sample meets or fails the BE testing criteria.

Examples

if (FALSE) {
nca_results <- system.file(package = "mbbe", "nca_results")
calc_power("c:/MBBErundir", 100, 0.05, "study", FALSE)
}