Check Requirements for Model-Based BE Assessment

This function verifies the necessary requirements for conducting a model-based bioequivalence (BE) assessment.

check_requirements(
  run_dir,
  samp_size,
  model_list,
  ngroups,
  reference_groups,
  test_groups,
  nmfe_path,
  use_check_identifiable,
  simulation_data_path,
  user_R_code,
  R_code_path = NULL
)

Arguments

run_dir

Character string specifying the directory containing the source model files.

samp_size

Integer specifying the number of samples.

model_list

List of character strings representing the model control files to be utilized for model averaging.

ngroups

Integer representing the number of groups for BE testing in the dataset.

reference_groups

Character vector indicating the groups classified as "reference".

test_groups

Character vector indicating the groups classified as "test".

nmfe_path

Character string indicating the path to the nmfe batch file (e.g., nmfe?.bat).

use_check_identifiable

Logical. Determines whether to verify the identifiability by checking if SADDLE_RESET is present in $EST.

simulation_data_path

Character string specifying the path to the data file used for simulation.

user_R_code

Logical. Should user-defined R code be executed post the NONMEM run for bootstrap?

R_code_path

Optional. Character string indicating the path to the user-defined R code to be executed post the NONMEM bootstrap. This returns a penalty value to be appended to the BIC.

Value

TRUE if all requirements pass and FALSE if any failures.

Details

The function checks for the following:

Presence of the .mod file containing ;;;;.*Start EST.

Verification that the sum of reference and test groups equals ngroups.

Absence of duplicate values in both reference and test groups.

Presence of the specified data file.

Validity of the provided nmfe_path.

(If requested) Verification for SADDLE_RESET in the context of identifiability.

Absence of repeat IDs in the dataset.

Availability of simulation data.