Calculate NCA Parameters from NONMEM Output

This function extracts and calculates Non-Compartmental Analysis (NCA) parameters (Cmax, AUCinf, AUClast) by group from a NONMEM simulation output file. The output file must adhere to certain specifications, including naming conventions and table header options.

getNCA(
  run_dir,
  this_sample,
  NumGroups,
  reference_groups,
  test_groups,
  NCA_end_time
)

Arguments

run_dir

Character string specifying the home directory where the NONMEM output files are located.

this_sample

Integer representing the specific sample being evaluated.

NumGroups

Integer specifying the total number of groups (e.g., 4 for an ABBA design).

reference_groups

Numeric vector indicating the group IDs that are designated as reference.

test_groups

Numeric vector indicating the group IDs that are designated as test.

NCA_end_time

Numeric value indicating the end time for the AUClast and AUCinf calculations. Calculation starts at TIME=0. A data point at TIME=0 is mandatory in the simulated study.

Value

Logical indicating the success or failure of the NCA calculation.

Details

The NONMEM output file should be present in the folder named MBBEsimN, where N is the simulation number. The $TABLE in the NONMEM file must use the ONEHEADER option and include specific columns such as ID, GROUP, SEQ, and PERIOD. For MBBE, the simulation control file, usually created by the write_sim function, will have the table specifications like: "$TABLE ID TIME GROUP PERIOD SEQ DV EVID NOPRINT NOAPPEND FILE=OUT.DAT ONEHEADER". This function reads the $TABLE output from the simulation (filename OUT.DAT) and performs NCA.

Examples

if (FALSE) {
getNCA("c:/runmodels", 1, 4, c(1, 2), c(3, 4), 72)
}