Execute MBBE Analysis

This function runs the MBBE analysis. It's typically called by run_mbbe_json which provides the necessary options via a JSON file.

run_mbbe(
  crash_value,
  ngroups,
  reference_groups,
  test_groups,
  num_parallel,
  samp_size,
  run_dir,
  model_source,
  nmfe_path,
  delta_parms,
  use_check_identifiable,
  NCA_end_time,
  rndseed,
  simulation_data_path,
  plan = c("multisession", "sequential", "multicore"),
  alpha_error = 0.05,
  NTID = FALSE,
  model_averaging_by = "study",
  user_R_code = FALSE,
  R_code_path = "",
  save_plots = FALSE,
  ...
)

Arguments

crash_value

Numeric. Value to be returned for BIC in models that crash during either bootstrap or simulation.

ngroups

Integer. Number of groups in the simulated data (e.g., an ABBA design has 4 groups).

reference_groups

Numeric vector. Indices of the groups representing the reference formulation (e.g., c(2,3) for an ABBA design).

test_groups

Numeric vector. Indices of the groups representing the test formulation (e.g., c(1,4) for an ABBA design).

num_parallel

Integer. Number of NONMEM processes (both bootstrap and simulation) to run concurrently.

samp_size

Integer. Size of the bootstrap and simulation samples.

run_dir

Character string. Directory for NONMEM execution.

model_source

Character string. Paths to the NONMEM control files for model averaging.

nmfe_path

Character string. Path to the nmfe executable.

delta_parms

Numeric. Parameter difference threshold defining identifiability.

use_check_identifiable

Logical. Should identifiability be checked based on the criterion defined by Aoki?

NCA_end_time

Numeric. The NCA calculation will start at 0 and end at this value.

rndseed

Integer. Random seed for reproducibility.

simulation_data_path

Character string. Path to the simulation dataset.

plan

Character string (default: "multisession"). Parallel execution plan. Can be "multisession", "sequential", or "multicore".

alpha_error

Numeric (default: 0.05). Alpha error rate for statistical tests.

NTID

Logical (default: FALSE). Is the drug a narrow therapeutic index drug?

model_averaging_by

Character string (default: "study"). Method of model averaging, either "study" or "subject".

user_R_code

Logical (default: FALSE). Should custom R code be used for model penalty?

R_code_path

Character string. If user_R_code is TRUE, this parameter defines the path to the custom R script.

save_plots

Logical (default: FALSE). Set to TRUE to save plot output.

...

Additional args

Value

A list containing:

Cmax_power: Power for Cmax

AUClast_power: Power for AUClast

AUCinf_power: Power for AUCinf

run_dir: Directory where the function was executed

Num_identifiable: Number of identifiable parameters

BICS: Bayesian Information Criterion Scores

Details

This function is primarily intended to be called by run_mbbe_json, which provides input parameters through a JSON configuration.